+++ types/adapter/entry-points/cli/projectConfig.d.ts
@@ -11,6 +11,7 @@ export type ConfigFile = {
     allowedIssueAuthors?: string;
     autoAssignManagerAuthors?: string;
     thresholdForAutoReject?: number;
+    thresholdForDispatchLoop?: number;
     workflowBlockerResolvedWebhookUrl?: string;
     projectName?: string;
     preparationProcessCheckCommand?: string;