+++ types/domain/usecases/isHumanComment.d.ts
@@ -1,0 +1,5 @@
+export declare const isHumanComment: (comment: {
+    author: string;
+    content: string;
+}, isTrustedAuthor: (author: string) => boolean) => boolean;
+//# sourceMappingURL=isHumanComment.d.ts.map